+++ src/main/java/com/example/banksimulation/SavingsAccount.java
@@ -4,7 +4,7 @@ public class SavingsAccount extends Account {
 
 
 
-    SavingsAccount(int accountNumber, int accountBalance, Customer accountHolder) {
+    SavingsAccount(int accountNumber, double accountBalance, Customer accountHolder) {
         super(accountNumber, accountBalance, accountHolder);
         }
